File Structure: Moved EventsInMannheim to views
