+++ src/views/EventsInMannheim/EventDisplay/index.jsx
