+++ src/views/EventsInMannheim/helper.js
